+++ tema 3/actor.h
@@ -11,14 +11,14 @@
 class actor : public persoana {
 protected:
     int nr_filme;
-    int suma;
+    double suma;
 
 public:
     explicit actor(std :: string CNP = "", std :: string nume = "", std :: string nume_film = "", int nr_filme = 0, double suma = 0);
     void citire(std :: istream& i) override;
     void print(std :: ostream& o) override;
 
-    int getSuma() const;
+    double getSuma() const;
 
     virtual ~actor();
 };
