+++ tema 3/contracte.h
@@ -10,10 +10,10 @@
 #include <tuple>
 
 class contracte {
-    std :: set < std :: tuple < persoana*, std :: string, int > > mapa;
+    std :: set < std :: tuple < persoana*, std :: string, double > > mapa;
 public:
     explicit contracte();
-    void insert(persoana* pers, int suma_standard);
+    void insert(persoana* pers, double suma_standard);
     friend std :: ostream& operator << (std :: ostream& o, contracte& c);
     ~contracte();
 };
